Added an opacity when the window is unfocused, and added opacity settings to the settingpanel

+++ Assets/Scripts/SaveAndLoad/LoadVolumeValue.cs
@@ -23,7 +23,7 @@ public class LoadVolumeValue : MonoBehaviour
 
         if (PlayerPrefs.HasKey("Volume"))
         {
-            int volume = PlayerPrefs.GetInt("Volume");
+            float volume = PlayerPrefs.GetFloat("Volume");
             slider.value = volume;
             changeSpriteOnValue.ChangeSprite(volume);
             audioMixer.SetFloat("Volume", volume);
